Add export helper functions
