+++ functions/index.js
@@ -4,7 +4,8 @@ const fs = require('fs-extra');
 const path = require('path');
 const os = require('os');
 const { Storage } = require('@google-cloud/storage');
-const { buildHeader, buildData } = require('./helper');
+const { buildMultitaskHeader, buildMultitaskData } = require('./helper_multitask');
+const { buildFocusHeader, buildFocusData } = require('./helper_testfocus');
 admin.initializeApp(functions.config().firebase);
 
 exports.csvJsonReport = functions.https.onRequest((request, response) => {
@@ -24,11 +25,41 @@ exports.csvJsonReport = functions.https.onRequest((request, response) => {
             querySnapshot.forEach(doc => {
                 dataList.push(doc.data())
             });
-            const header = buildHeader();
-            const dataRow = buildData(dataList);
+            const header = buildMultitaskHeader();
+            const dataRow = buildMultitaskData(dataList);
             return header.concat(dataRow);
         }).then(csv => {
             return fs.outputFile(tempFilePath, csv);
+        }).then(() => {
+            return myBucket.upload(tempFilePath, { destination: fileName })
+        }).then(() => {
+            response.redirect('https://console.firebase.google.com/');
+        }).catch(err => console.log(err))
+});
+
+exports.csvJsonReport2 = functions.https.onRequest((request, response) => {
+    response.set('Access-Control-Allow-Origin', '*');
+    const db = admin.firestore();
+    const bucketName = 'common-research.appspot.com';
+    const fileName = 'reports/report2.csv';
+    const tempFilePath = path.join(os.tmpdir(), fileName);
+
+    const storage = new Storage();
+    const myBucket = storage.bucket(bucketName);
+
+    db.collection('testfocus')
+        .get()
+        .then(querySnapshot => {
+            const dataList = []
+            querySnapshot.forEach(doc => {
+                dataList.push(doc.data())
+            });
+            const header = buildFocusHeader(); // Modify this with new helper methods
+            const dataRow = buildFocusData(dataList); // And this
+            return header.concat(dataRow);
+        }).then(csv => {
+            return fs.outputFile(tempFilePath, csv);
+
         }).then(() => {
             return myBucket.upload(tempFilePath, { destination: fileName })
         }).then(() => {
